Split runtime and provider tests
