+++ index.test.ts
@@ -3,172 +3,12 @@ import { tmpdir } from "node:os";
 import { join } from "node:path";
 import { describe, expect, it, vi } from "vitest";
 import plugin from "./index.js";
+import {
+  getRegisteredProvider,
+  getRegisteredProviderWithAuth,
+} from "./provider/test-harness.js";
 
 describe("nanogpt plugin entry", () => {
-  function getRegisteredProviderHarness(overrideConfig: Record<string, unknown> = {}) {
-    const providers: unknown[] = [];
-    const warn = vi.fn();
-    const info = vi.fn();
-    plugin.register(
-      {
-        pluginConfig: { enableRepair: false, ...overrideConfig },
-        runtime: {
-          logging: {
-            shouldLogVerbose() {
-              return false;
-            },
-          },
-        },
-        logger: {
-          warn,
-          info,
-        },
-        registerProvider(provider: unknown) {
-          providers.push(provider);
-        },
-        registerWebSearchProvider() {},
-        registerImageGenerationProvider() {},
-      } as never,
-    );
-
-    return {
-      warn,
-      info,
-      provider: providers[0] as {
-        applyNativeStreamingUsageCompat?: (ctx: {
-          providerConfig: {
-            api: string;
-            baseUrl?: string;
-            models?: Array<Record<string, unknown>>;
-          };
-        }) => unknown;
-        wrapStreamFn?: (ctx: {
-          streamFn?: (...args: unknown[]) => unknown;
-          modelId: string;
-          model?: {
-            id?: string;
-          };
-        }) => unknown;
-        normalizeToolSchemas?: (ctx: {
-          provider: string;
-          modelId?: string;
-          model?: {
-            id: string;
-            provider?: string;
-            api?: string;
-            baseUrl?: string;
-          };
-          tools: Array<Record<string, unknown>>;
-        }) => unknown;
-        inspectToolSchemas?: (ctx: {
-          provider: string;
-          modelId?: string;
-          model?: {
-            id: string;
-            provider?: string;
-            api?: string;
-            baseUrl?: string;
-          };
-          tools: Array<Record<string, unknown>>;
-        }) => unknown;
-        augmentModelCatalog?: (ctx: {
-          agentDir?: string;
-          config?: Record<string, unknown>;
-          env?: Record<string, string | undefined>;
-          entries: Array<Record<string, unknown>>;
-        }) => unknown;
-        normalizeResolvedModel?: (ctx: {
-          agentDir?: string;
-          provider: string;
-          modelId: string;
-          model: {
-            id: string;
-            name: string;
-            provider: string;
-            api: string;
-            baseUrl?: string;
-            reasoning: boolean;
-            input: Array<"text" | "image" | "document">;
-            cost: {
-              input: number;
-              output: number;
-              cacheRead: number;
-              cacheWrite: number;
-            };
-            contextWindow: number;
-            maxTokens: number;
-            compat?: Record<string, unknown>;
-          };
-        }) => unknown;
-        resolveDynamicModel?: (ctx: {
-          agentDir?: string;
-          env?: Record<string, string | undefined>;
-          provider: string;
-          modelId: string;
-          modelRegistry: unknown;
-          providerConfig?: {
-            api?: string;
-            baseUrl?: string;
-            models?: Array<Record<string, unknown>>;
-          };
-        }) => unknown;
-        resolveUsageAuth?: unknown;
-        fetchUsageSnapshot?: unknown;
-        matchesContextOverflowError?: (ctx: {
-          provider?: string;
-          modelId?: string;
-          errorMessage: string;
-        }) => boolean | undefined;
-        classifyFailoverReason?: (ctx: {
-          provider?: string;
-          modelId?: string;
-          errorMessage: string;
-        }) => string | null | undefined;
-      },
-    };
-  }
-
-  function getRegisteredProvider(overrideConfig: Record<string, unknown> = {}) {
-    return getRegisteredProviderHarness(overrideConfig).provider;
-  }
-
-  function getRegisteredProviderWithAuth() {
-    const providers: unknown[] = [];
-    plugin.register(
-      {
-        pluginConfig: {},
-        runtime: {
-          logging: {
-            shouldLogVerbose() {
-              return false;
-            },
-          },
-        },
-        logger: {
-          warn: vi.fn(),
-          info: vi.fn(),
-        },
-        registerProvider(provider: unknown) {
-          providers.push(provider);
-        },
-        registerWebSearchProvider() {},
-        registerImageGenerationProvider() {},
-      } as never,
-    );
-    return providers[0] as {
-      auth?: Array<{
-        runNonInteractive?: (ctx: {
-          opts?: Record<string, unknown>;
-          config: Record<string, unknown>;
-          env: Record<string, string | undefined>;
-          agentDir: string;
-          resolveApiKey: () => Promise<{ source: string } | null>;
-          toApiKeyCredential: () => unknown;
-        }) => Promise<Record<string, unknown> | null>;
-      }>;
-    };
-  }
-
   it("exports the expected plugin metadata", () => {
     expect(plugin.id).toBe("nanogpt");
     expect(plugin.name).toBe("NanoGPT Provider");
@@ -217,125 +57,6 @@ describe("nanogpt plugin entry", () => {
     );
     expect((providers[0] as { applyNativeStreamingUsageCompat?: unknown }).applyNativeStreamingUsageCompat).toEqual(
       expect.any(Function),
-    );
-  });
-
-  it("classifies structured NanoGPT rate limits and logs the mapped reason once", () => {
-    const { provider, warn } = getRegisteredProviderHarness();
-    expect(provider.classifyFailoverReason).toEqual(expect.any(Function));
-
-    const errorMessage = JSON.stringify({
-      error: {
-        message: "Daily request limit exceeded",
-        type: "rate_limit_error",
-        code: "daily_rpd_limit_exceeded",
-      },
-      status: 429,
-    });
-
-    expect(
-      provider.classifyFailoverReason?.({
-        provider: "nanogpt",
-        modelId: "moonshotai/kimi-k2.5:thinking",
-        errorMessage,
-      }),
-    ).toBe("rate_limit");
-
-    expect(
-      provider.classifyFailoverReason?.({
-        provider: "nanogpt",
-        modelId: "moonshotai/kimi-k2.5:thinking",
-        errorMessage,
-      }),
-    ).toBe("rate_limit");
-
-    expect(warn).toHaveBeenCalledTimes(1);
-    expect(warn).toHaveBeenCalledWith(
-      expect.stringContaining("classified as rate_limit"),
-    );
-  });
-
-  it("warns and falls through when NanoGPT returns a recognized but unmapped error code", () => {
-    const { provider, warn } = getRegisteredProviderHarness();
-    expect(provider.classifyFailoverReason).toEqual(expect.any(Function));
-
-    expect(
-      provider.classifyFailoverReason?.({
-        provider: "nanogpt",
-        modelId: "moonshotai/kimi-k2.5:thinking",
-        errorMessage: JSON.stringify({
-          error: {
-            message: "Every configured fallback failed",
-            type: "server_error",
-            code: "all_fallbacks_failed",
-          },
-          status: 409,
-        }),
-      }),
-    ).toBeUndefined();
-
-    expect(warn).toHaveBeenCalledTimes(1);
-    expect(warn).toHaveBeenCalledWith(
-      expect.stringContaining("recognized but not mapped"),
-    );
-  });
-
-  it("warns and falls through when NanoGPT returns an unknown structured error envelope", () => {
-    const { provider, warn } = getRegisteredProviderHarness();
-    expect(provider.classifyFailoverReason).toEqual(expect.any(Function));
-
-    expect(
-      provider.classifyFailoverReason?.({
-        provider: "nanogpt",
-        modelId: "moonshotai/kimi-k2.5:thinking",
-        errorMessage: JSON.stringify({
-          error: {
-            detail: "surprising payload",
-          },
-          status: 418,
-        }),
-      }),
-    ).toBeUndefined();
-
-    expect(warn).toHaveBeenCalledTimes(1);
-    expect(warn).toHaveBeenCalledWith(
-      expect.stringContaining("Unknown NanoGPT API error envelope"),
-    );
-  });
-
-  it("routes NanoGPT context length errors through the context overflow hook", () => {
-    const { provider, warn } = getRegisteredProviderHarness();
-    expect(provider.matchesContextOverflowError).toEqual(expect.any(Function));
-    expect(provider.classifyFailoverReason).toEqual(expect.any(Function));
-
-    const errorMessage = JSON.stringify({
-      error: {
-        message: "Context length exceeded",
-        type: "invalid_request_error",
-        code: "context_length_exceeded",
-      },
-      status: 400,
-    });
-
-    expect(
-      provider.matchesContextOverflowError?.({
-        provider: "nanogpt",
-        modelId: "moonshotai/kimi-k2.5:thinking",
-        errorMessage,
-      }),
-    ).toBe(true);
-
-    expect(
-      provider.classifyFailoverReason?.({
-        provider: "nanogpt",
-        modelId: "moonshotai/kimi-k2.5:thinking",
-        errorMessage,
-      }),
-    ).toBeUndefined();
-
-    expect(warn).toHaveBeenCalledTimes(1);
-    expect(warn).toHaveBeenCalledWith(
-      expect.stringContaining("context overflow handling"),
     );
   });
 
@@ -446,140 +167,6 @@ describe("nanogpt plugin entry", () => {
       model: { id: "moonshotai/kimi-k2.5:thinking" },
     });
     expect(wrappedFn).toBe(baseStreamFn);
-  });
-
-  it("keeps web_fetch untouched on Kimi models because aliasing is disabled", () => {
-    const provider = getRegisteredProvider();
-    const normalizeToolSchemas = provider.normalizeToolSchemas;
-    expect(normalizeToolSchemas).toEqual(expect.any(Function));
-
-    const fetchTool = {
-      name: "web_fetch",
-      description: "Fetch and extract readable content from a URL",
-      parameters: { type: "object" },
-      execute: async () => ({ ok: true }),
-    };
-
-    const normalized = normalizeToolSchemas?.({
-      provider: "nanogpt",
-      modelId: "moonshotai/kimi-k2.5:thinking",
-      model: {
-        id: "moonshotai/kimi-k2.5:thinking",
-        provider: "nanogpt",
-        api: "openai-completions",
-      },
-      tools: [fetchTool],
-    }) as Array<{ name: string }> | null;
-
-    expect(normalized).toBeNull();
-  });
-
-  it("adds GLM schema hints without renaming web_fetch", () => {
-    const provider = getRegisteredProvider();
-    const normalizeToolSchemas = provider.normalizeToolSchemas;
-    expect(normalizeToolSchemas).toEqual(expect.any(Function));
-
-    const fetchTool = {
-      name: "web_fetch",
-      description: "Fetch and extract readable content from a URL",
-      parameters: { type: "object" },
-      execute: async () => ({ ok: true }),
-    };
-
-    const normalized = normalizeToolSchemas?.({
-      provider: "nanogpt",
-      modelId: "zai-org/glm-5:thinking",
-      model: {
-        id: "zai-org/glm-5:thinking",
-        provider: "nanogpt",
-        api: "openai-completions",
-      },
-      tools: [fetchTool],
-    }) as Array<{ name: string; description?: string }> | null;
-
-    expect(normalized).toHaveLength(1);
-    expect(normalized?.[0]).toMatchObject({
-      name: "web_fetch",
-      description: expect.stringContaining("NanoGPT GLM tip:"),
-    });
-    expect(normalized?.[0]?.name).not.toBe("fetch_web_page");
-    expect(normalized?.[0]?.description).toContain(
-      "include required ref/selector/fields arguments explicitly when the tool needs them.",
-    );
-    expect(fetchTool.description).toBe("Fetch and extract readable content from a URL");
-  });
-
-  it("adds Qwen schema hints that steer models away from leaked XML-like tool text", () => {
-    const provider = getRegisteredProvider();
-    const normalizeToolSchemas = provider.normalizeToolSchemas;
-    expect(normalizeToolSchemas).toEqual(expect.any(Function));
-
-    const execTool = {
-      name: "exec",
-      description: "Execute a shell command",
-      parameters: {
-        type: "object",
-        properties: {
-          command: { type: "string" },
-        },
-        required: ["command"],
-      },
-      execute: async () => ({ ok: true }),
-    };
-
-    const normalized = normalizeToolSchemas?.({
-      provider: "nanogpt",
-      modelId: "qwen/Qwen3.6-35B-A3B:thinking",
-      model: {
-        id: "qwen/Qwen3.6-35B-A3B:thinking",
-        provider: "nanogpt",
-        api: "openai-completions",
-      },
-      tools: [execTool],
-    }) as Array<{ name: string; description?: string }> | null;
-
-    expect(normalized).toHaveLength(1);
-    expect(normalized?.[0]).toMatchObject({
-      name: "exec",
-      description: expect.stringContaining("NanoGPT Qwen tip:"),
-    });
-    expect(normalized?.[0]?.description).toContain('{"command":"..."}');
-    expect(normalized?.[0]?.description).toContain("<exec>...</exec>");
-    expect(execTool.description).toBe("Execute a shell command");
-  });
-
-  it("surfaces Qwen schema diagnostics for tools that are hard to revalidate", () => {
-    const provider = getRegisteredProvider();
-    const inspectToolSchemas = provider.inspectToolSchemas;
-    expect(inspectToolSchemas).toEqual(expect.any(Function));
-
-    const diagnostics = inspectToolSchemas?.({
-      provider: "nanogpt",
-      modelId: "qwen/Qwen3.6-35B-A3B",
-      model: {
-        id: "qwen/Qwen3.6-35B-A3B",
-        provider: "nanogpt",
-        api: "openai-completions",
-      },
-      tools: [
-        {
-          name: "exec",
-          parameters: { type: "object" },
-          execute: async () => ({ ok: true }),
-        },
-      ],
-    }) as Array<{ toolName: string; toolIndex?: number; violations: string[] }> | null;
-
-    expect(diagnostics).toEqual([
-      {
-        toolName: "exec",
-        toolIndex: 0,
-        violations: expect.arrayContaining([
-          expect.stringContaining("missing description"),
-          expect.stringContaining("no named properties"),
-        ]),
-      },
-    ]);
   });
 
   it("surfaces discovered NanoGPT models from models.json into catalog augmentation", () => {
