+++ runtime.test.ts
@@ -1,37 +1,13 @@
-import {
-  NANOGPT_BASE_URL,
-  NANOGPT_SUBSCRIPTION_BASE_URL,
-  NANOGPT_PAID_BASE_URL,
-  NANOGPT_PERSONALIZED_BASE_URL,
-} from "./models.js";
 import { afterEach, describe, expect, it, vi } from "vitest";
-import { sanitizeApiKey } from "./shared/http.js";
-import {
-  buildNanoGptRequestHeaders,
-  probeNanoGptSubscription,
-  resolveCatalogBaseUrl,
-  resolveCatalogSource,
-  resolveNanoGptRoutingMode,
-  resolveRequestBaseUrl,
-} from "./runtime/routing.js";
+import { resolveNanoGptRoutingMode } from "./runtime/routing.js";
 import { resetNanoGptRuntimeState } from "./runtime.js";
 import { discoverNanoGptModels } from "./runtime/discovery.js";
 import { getNanoGptConfig, resolveNanoGptRequestApi } from "./runtime/config.js";
 import { resolveNanoGptDynamicModel } from "./runtime/dynamic-models.js";
-import { fetchNanoGptSelectedProviderPricing } from "./runtime/provider-pricing.js";
-import { fetchNanoGptUsageSnapshot, resolveNanoGptUsageAuth } from "./runtime/usage.js";
 
 afterEach(() => {
   resetNanoGptRuntimeState();
   vi.unstubAllGlobals();
-});
-
-describe("sanitizeApiKey", () => {
-  it("removes carriage returns and line feeds to prevent HTTP header injection", () => {
-    expect(sanitizeApiKey("test-key")).toBe("test-key");
-    expect(sanitizeApiKey("test-key\r\nInjected: true")).toBe("test-keyInjected: true");
-    expect(sanitizeApiKey("\ntest\r")).toBe("test");
-  });
 });
 
 describe("getNanoGptConfig", () => {
@@ -60,297 +36,6 @@ describe("resolveNanoGptRequestApi", () => {
 
   it("keeps explicit responses requestApi", () => {
     expect(resolveNanoGptRequestApi({ requestApi: "responses" })).toBe("openai-responses");
-  });
-});
-
-describe("resolveNanoGptRoutingMode", () => {
-  it("returns explicit paygo without probing", async () => {
-    const fetchSpy = vi.fn();
-    vi.stubGlobal("fetch", fetchSpy);
-
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "paygo" },
-        apiKey: "test-key",
-      }),
-    ).resolves.toBe("paygo");
-
-    expect(fetchSpy).not.toHaveBeenCalled();
-  });
-
-  it("treats a state-only active usage payload as subscription", async () => {
-    const fetchSpy = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ state: "active" }),
-    });
-    vi.stubGlobal("fetch", fetchSpy);
-
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "auto" },
-        apiKey: "state-only-key",
-      }),
-    ).resolves.toBe("subscription");
-
-    expect(fetchSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it("treats a future grace period as subscription", async () => {
-    const fetchSpy = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({
-        state: "grace",
-        active: false,
-        graceUntil: Date.now() + 60_000,
-      }),
-    });
-    vi.stubGlobal("fetch", fetchSpy);
-
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "auto" },
-        apiKey: "grace-key",
-      }),
-    ).resolves.toBe("subscription");
-
-    expect(fetchSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it("favors subscription when the usage probe errors in auto mode", async () => {
-    const fetchSpy = vi.fn().mockRejectedValue(new Error("usage probe failed"));
-    vi.stubGlobal("fetch", fetchSpy);
-
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "auto" },
-        apiKey: "probe-error-key",
-      }),
-    ).resolves.toBe("subscription");
-
-    expect(fetchSpy).toHaveBeenCalledTimes(1);
-  });
-
-  it("does not cache a probe failure as paygo for the next auto-routed call", async () => {
-    const fetchSpy = vi
-      .fn()
-      .mockRejectedValueOnce(new Error("usage probe failed"))
-      .mockResolvedValueOnce({
-        ok: true,
-        json: async () => ({ subscribed: true }),
-      });
-    vi.stubGlobal("fetch", fetchSpy);
-
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "auto" },
-        apiKey: "retry-key",
-      }),
-    ).resolves.toBe("subscription");
-
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "auto" },
-        apiKey: "retry-key",
-      }),
-    ).resolves.toBe("subscription");
-
-    expect(fetchSpy).toHaveBeenCalledTimes(2);
-  });
-
-  it("caches subscription status per api key", async () => {
-    const fetchSpy = vi
-      .fn()
-      .mockResolvedValueOnce({
-        ok: true,
-        json: async () => ({ subscribed: true }),
-      })
-      .mockResolvedValueOnce({
-        ok: true,
-        json: async () => ({ subscribed: false }),
-      });
-    vi.stubGlobal("fetch", fetchSpy);
-
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "auto" },
-        apiKey: "key-a",
-      }),
-    ).resolves.toBe("subscription");
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "auto" },
-        apiKey: "key-a",
-      }),
-    ).resolves.toBe("subscription");
-    await expect(
-      resolveNanoGptRoutingMode({
-        config: { routingMode: "auto" },
-        apiKey: "key-b",
-      }),
-    ).resolves.toBe("paygo");
-
-    expect(fetchSpy).toHaveBeenCalledTimes(2);
-  });
-});
-
-describe("resolveCatalogSource", () => {
-  it("maps auto to subscription when routing resolved to subscription", () => {
-    expect(resolveCatalogSource({ config: {}, routingMode: "subscription" })).toBe("subscription");
-  });
-
-  it("maps auto to canonical when routing resolved to paygo", () => {
-    expect(resolveCatalogSource({ config: {}, routingMode: "paygo" })).toBe("canonical");
-  });
-});
-
-describe("resolveRequestBaseUrl", () => {
-  it("uses the base API for responses requests even when routing resolved to subscription", () => {
-    expect(
-      resolveRequestBaseUrl({
-        config: { requestApi: "responses" },
-        routingMode: "subscription",
-      }),
-    ).toBe("https://nano-gpt.com/api/v1");
-  });
-
-  it("keeps the subscription API for completions requests on subscription routing", () => {
-    expect(
-      resolveRequestBaseUrl({
-        config: { requestApi: "completions" },
-        routingMode: "subscription",
-      }),
-    ).toBe("https://nano-gpt.com/api/subscription/v1");
-  });
-});
-
-describe("buildNanoGptRequestHeaders", () => {
-  it("adds provider override headers for paygo routing", () => {
-    expect(
-      buildNanoGptRequestHeaders({
-        apiKey: "test-key",
-        config: { provider: "openrouter" },
-        routingMode: "paygo",
-      }),
-    ).toEqual({
-      Authorization: "Bearer test-key",
-      "X-Provider": "openrouter",
-    });
-  });
-
-  it("sanitizes provider header values before sending them", () => {
-    expect(
-      buildNanoGptRequestHeaders({
-        apiKey: "test-key\r\nInjected: true",
-        config: { provider: "openrouter\r\nInjected: true" },
-        routingMode: "paygo",
-      }),
-    ).toEqual({
-      Authorization: "Bearer test-keyInjected: true",
-      "X-Provider": "openrouterInjected: true",
-    });
-  });
-
-  it("ignores provider override during subscription routing", () => {
-    expect(
-      buildNanoGptRequestHeaders({
-        apiKey: "test-key",
-        config: { provider: "openrouter" },
-        routingMode: "subscription",
-      }),
-    ).toEqual({
-      Authorization: "Bearer test-key",
-    });
-  });
-
-  it("does not add extra headers when no provider override is configured", () => {
-    expect(
-      buildNanoGptRequestHeaders({
-        apiKey: "test-key",
-        config: {},
-        routingMode: "subscription",
-      }),
-    ).toEqual({
-      Authorization: "Bearer test-key",
-    });
-  });
-});
-
-describe("resolveNanoGptUsageAuth", () => {
-  it("resolves the NanoGPT API key from config/store/env", async () => {
-    await expect(
-      resolveNanoGptUsageAuth({
-        env: { NANOGPT_API_KEY: "env-key" },
-        resolveApiKeyFromConfigAndStore: () => "stored-key",
-      } as never),
-    ).resolves.toEqual({ token: "stored-key" });
-  });
-});
-
-describe("fetchNanoGptUsageSnapshot", () => {
-  it("maps NanoGPT quota windows into OpenClaw usage windows", async () => {
-    const fetchSpy = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({
-        active: true,
-        state: "active",
-        daily: {
-          used: 5,
-          remaining: 1995,
-          percentUsed: 0.0025,
-          resetAt: 1738540800000,
-        },
-        monthly: {
-          used: 45,
-          remaining: 59955,
-          percentUsed: 0.00075,
-          resetAt: 1739404800000,
-        },
-      }),
-    });
-
-    await expect(
-      fetchNanoGptUsageSnapshot({
-        token: "test-key",
-        timeoutMs: 1_000,
-        fetchFn: fetchSpy,
-      } as never),
-    ).resolves.toMatchObject({
-      provider: "nanogpt",
-      displayName: "NanoGPT",
-      plan: "active",
-      windows: [
-        {
-          label: "Daily",
-          usedPercent: 0.25,
-          resetAt: 1738540800000,
-        },
-        {
-          label: "Monthly",
-          usedPercent: 0.075,
-          resetAt: 1739404800000,
-        },
-      ],
-    });
-  });
-
-  it("returns an error snapshot on non-200 responses", async () => {
-    const fetchSpy = vi.fn().mockResolvedValue(
-      new Response("Unauthorized", {
-        status: 401,
-        headers: { "Content-Type": "text/plain" },
-      }),
-    );
-
-    await expect(
-      fetchNanoGptUsageSnapshot({
-        token: "bad-key",
-        timeoutMs: 1_000,
-        fetchFn: fetchSpy,
-      } as never),
-    ).resolves.toMatchObject({
-      provider: "nanogpt",
-      error: expect.stringContaining("HTTP 401"),
-    });
   });
 });
 
@@ -775,164 +460,14 @@ describe("resetNanoGptRuntimeState", () => {
       config: { routingMode: "auto" },
       apiKey: "test-clearing",
     });
-    
+
     resetNanoGptRuntimeState();
-    
+
     await resolveNanoGptRoutingMode({
       config: { routingMode: "auto" },
       apiKey: "test-clearing",
     });
 
     expect(fetchSpy).toHaveBeenCalledTimes(2);
-  });
-
-  it("clears the provider pricing cache", async () => {
-    const fetchSpy = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({
-        supportsProviderSelection: true,
-        providers: [
-          {
-            provider: "openrouter",
-            available: true,
-            pricing: {
-              inputPer1kTokens: 0.00042,
-              outputPer1kTokens: 0.0018375,
-              unit: "per_1k_tokens",
-            },
-          },
-        ],
-      }),
-    });
-    vi.stubGlobal("fetch", fetchSpy);
-
-    const params = {
-      apiKey: "test-key",
-      modelId: "test-model",
-      provider: "openrouter",
-    };
-
-    await fetchNanoGptSelectedProviderPricing(params);
-    await fetchNanoGptSelectedProviderPricing(params);
-    expect(fetchSpy).toHaveBeenCalledTimes(1);
-
-    resetNanoGptRuntimeState();
-
-    await fetchNanoGptSelectedProviderPricing(params);
-    expect(fetchSpy).toHaveBeenCalledTimes(2);
-  });
-});
-
-describe("fetchNanoGptSelectedProviderPricing", () => {
-  it("returns null if provider normalizes to empty", async () => {
-    const result = await fetchNanoGptSelectedProviderPricing({
-      apiKey: "test-key",
-      modelId: "test-model",
-      provider: "   ", // normalizes to empty string
-    });
-    expect(result).toBeNull();
-  });
-
-  it("returns null if fetch throws an error", async () => {
-    vi.stubGlobal("fetch", vi.fn().mockRejectedValue(new Error("Network error")));
-    const result = await fetchNanoGptSelectedProviderPricing({
-      apiKey: "test-key",
-      modelId: "test-model",
-      provider: "valid-provider",
-    });
-    expect(result).toBeNull();
-  });
-
-  it("returns null if fetch response is not ok", async () => {
-    vi.stubGlobal("fetch", vi.fn().mockResolvedValue({ ok: false }));
-    const result = await fetchNanoGptSelectedProviderPricing({
-      apiKey: "test-key",
-      modelId: "test-model",
-      provider: "valid-provider",
-    });
-    expect(result).toBeNull();
-  });
-
-  it("returns null if payload is invalid or does not support provider selection", async () => {
-    const fetchSpy = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({ supportsProviderSelection: false }),
-    });
-    vi.stubGlobal("fetch", fetchSpy);
-    const result = await fetchNanoGptSelectedProviderPricing({
-      apiKey: "test-key",
-      modelId: "test-model",
-      provider: "valid-provider",
-    });
-    expect(result).toBeNull();
-  });
-
-  it("shares an in-flight provider pricing lookup across concurrent calls", async () => {
-    const fetchSpy = vi.fn().mockResolvedValue({
-      ok: true,
-      json: async () => ({
-        supportsProviderSelection: true,
-        providers: [
-          {
-            provider: "openrouter",
-            available: true,
-            pricing: {
-              inputPer1kTokens: 0.00042,
-              outputPer1kTokens: 0.0018375,
-              unit: "per_1k_tokens",
-            },
-          },
-        ],
-      }),
-    });
-    vi.stubGlobal("fetch", fetchSpy);
-
-    const params = {
-      apiKey: "test-key",
-      modelId: "test-model",
-      provider: "openrouter",
-    };
-
-    const [first, second] = await Promise.all([
-      fetchNanoGptSelectedProviderPricing(params),
-      fetchNanoGptSelectedProviderPricing(params),
-    ]);
-
-    expect(first).toEqual(second);
-    expect(first).toMatchObject({
-      inputPer1kTokens: 0.00042,
-      outputPer1kTokens: 0.0018375,
-      unit: "per_1k_tokens",
-    });
-    expect(fetchSpy).toHaveBeenCalledTimes(1);
-
-    await expect(fetchNanoGptSelectedProviderPricing(params)).resolves.toMatchObject({
-      inputPer1kTokens: 0.00042,
-      outputPer1kTokens: 0.0018375,
-      unit: "per_1k_tokens",
-    });
-    expect(fetchSpy).toHaveBeenCalledTimes(1);
-  });
-});
-
-describe("resolveCatalogBaseUrl", () => {
-  it("maps subscription source to the subscription base URL", () => {
-    expect(resolveCatalogBaseUrl("subscription")).toBe(NANOGPT_SUBSCRIPTION_BASE_URL);
-  });
-
-  it("maps paid source to the paid base URL", () => {
-    expect(resolveCatalogBaseUrl("paid")).toBe(NANOGPT_PAID_BASE_URL);
-  });
-
-  it("maps personalized source to the personalized base URL", () => {
-    expect(resolveCatalogBaseUrl("personalized")).toBe(NANOGPT_PERSONALIZED_BASE_URL);
-  });
-
-  it("maps canonical source to the standard base URL", () => {
-    expect(resolveCatalogBaseUrl("canonical")).toBe(NANOGPT_BASE_URL);
-  });
-
-  it("maps an unknown/default source to the standard base URL", () => {
-    expect(resolveCatalogBaseUrl("unknown" as any)).toBe(NANOGPT_BASE_URL);
   });
 });
